implementation/requirements.txt
Django==5.2.4
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
pymongo==4.15.5
cryptography==46.0.3
google-genai
scikit-learn==1.7.0
numpy==2.2.6
pandas==2.3.1
joblib==1.5.1
transformers==4.47.1
sentence-transformers==3.3.1
torch>=2.0.0
